kdgebjholj

+++ GameClient/Projectile.h
@@ -10,9 +10,9 @@ class Projectile : public ServerEntity
 private:
 	FLOAT3 m_direction;
 	float m_movementSpeed;
-	int m_lifetime;
+	float m_lifetime;
 public:
-	Projectile(FLOAT3 _position, FLOAT3 _direction, int _lifetime, float _movementSpeed);
+	Projectile(FLOAT3 _position, FLOAT3 _direction, float _lifetime, float _movementSpeed);
 	virtual ~Projectile();
 
 	void update(float dt);
